+++ src/Symfony/Component/Messenger/DataCollector/MessengerDataCollector.php
@@ -98,14 +98,6 @@ class MessengerDataCollector extends DataCollector implements LateDataCollectorI
             'caller' => $tracedMessage['caller'],
         );
 
-        if (array_key_exists('result', $tracedMessage)) {
-            $result = $tracedMessage['result'];
-            $debugRepresentation['result'] = array(
-                'type' => \is_object($result) ? \get_class($result) : \gettype($result),
-                'value' => $result,
-            );
-        }
-
         if (isset($tracedMessage['exception'])) {
             $exception = $tracedMessage['exception'];