+++ src/Symfony/Component/Messenger/Tests/DataCollector/MessengerDataCollectorTest.php
@@ -33,15 +33,12 @@ class MessengerDataCollectorTest extends TestCase
         $this->dumper->setColors(false);
     }
 
-    /**
-     * @dataProvider getHandleTestData
-     */
-    public function testHandle($returnedValue, $expected)
+    public function testHandle()
     {
         $message = new DummyMessage('dummy message');
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
-        $bus->method('dispatch')->with($message)->willReturn($returnedValue);
+        $bus->method('dispatch')->with($message);
         $bus = new TraceableMessageBus($bus);
 
         $collector = new MessengerDataCollector();
@@ -54,13 +51,9 @@ class MessengerDataCollectorTest extends TestCase
         $messages = $collector->getMessages();
         $this->assertCount(1, $messages);
 
-        $this->assertStringMatchesFormat($expected, $this->getDataAsString($messages[0]));
-    }
-
-    public function getHandleTestData()
-    {
         $file = __FILE__;
-        $messageDump = <<<DUMP
+        $expected = <<<DUMP
+array:4 [
   "bus" => "default"
   "envelopeItems" => null
   "message" => array:2 [
@@ -74,48 +67,10 @@ class MessengerDataCollectorTest extends TestCase
     "file" => "$file"
     "line" => %d
   ]
+]
 DUMP;
 
-        yield 'no returned value' => array(
-            null,
-            <<<DUMP
-array:5 [
-$messageDump
-  "result" => array:2 [
-    "type" => "NULL"
-    "value" => null
-  ]
-]
-DUMP
-        );
-
-        yield 'scalar returned value' => array(
-            'returned value',
-            <<<DUMP
-array:5 [
-$messageDump
-  "result" => array:2 [
-    "type" => "string"
-    "value" => "returned value"
-  ]
-]
-DUMP
-        );
-
-        yield 'array returned value' => array(
-            array('returned value'),
-            <<<DUMP
-array:5 [
-$messageDump
-  "result" => array:2 [
-    "type" => "array"
-    "value" => array:1 [
-      0 => "returned value"
-    ]
-  ]
-]
-DUMP
-        );
+        $this->assertStringMatchesFormat($expected, $this->getDataAsString($messages[0]));
     }
 
     public function testHandleWithException()