+++ src/Symfony/Component/Messenger/Tests/Handler/ChainHandlerTest.php
@@ -17,7 +17,7 @@ use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 
 class ChainHandlerTest extends TestCase
 {
-    public function testItCallsTheHandlersAndReturnsAllResults()
+    public function testItCallsTheHandlers()
     {
         $message = new DummyMessage('Hey');
 
@@ -26,19 +26,15 @@ class ChainHandlerTest extends TestCase
             ->expects($this->once())
             ->method('__invoke')
             ->with($message)
-            ->willReturn('Hello')
         ;
         $handler2 = $this->createPartialMock(\stdClass::class, array('__invoke'));
         $handler2
             ->expects($this->once())
             ->method('__invoke')
             ->with($message)
-            ->willReturn('World')
         ;
 
-        $results = (new ChainHandler(array($handler1, $handler2)))($message);
-
-        $this->assertSame(array('Hello', 'World'), $results);
+        (new ChainHandler(array($handler1, $handler2)))($message);
     }
 
     /**